Completed initial connection to server
