+++ website/.cache/148b8d5d5a9ae85d151cdf7bdc5ac088.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/fbjs/lib/camelizeStyleName.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/package.json","includedInParent":true,"mtime":1535426282286},{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"./camelize","loc":{"line":12,"column":23},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/camelizeStyleName.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/camelize.js"}],"generated":{"js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\n'use strict';\n\nvar camelize = require('./camelize');\n\nvar msPattern = /^-ms-/;\n\n/**\n * Camelcases a hyphenated CSS property name, for example:\n *\n *   > camelizeStyleName('background-color')\n *   < \"backgroundColor\"\n *   > camelizeStyleName('-moz-transition')\n *   < \"MozTransition\"\n *   > camelizeStyleName('-ms-transition')\n *   < \"msTransition\"\n *\n * As Andi Smith suggests\n * (http://www.andismith.com/blog/2012/02/modernizr-prefixed/), an `-ms` prefix\n * is converted to lowercase `ms`.\n *\n * @param {string} string\n * @return {string}\n */\nfunction camelizeStyleName(string) {\n  return camelize(string.replace(msPattern, 'ms-'));\n}\n\nmodule.exports = camelizeStyleName;","map":{"mappings":[{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":22,"column":0},"generated":{"line":22,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":23,"column":0},"generated":{"line":23,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":24,"column":0},"generated":{"line":24,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":25,"column":0},"generated":{"line":25,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":26,"column":0},"generated":{"line":26,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":27,"column":0},"generated":{"line":27,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":28,"column":0},"generated":{"line":28,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":29,"column":0},"generated":{"line":29,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":30,"column":0},"generated":{"line":30,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":31,"column":0},"generated":{"line":31,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":32,"column":0},"generated":{"line":32,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":33,"column":0},"generated":{"line":33,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":34,"column":0},"generated":{"line":34,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":35,"column":0},"generated":{"line":35,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":36,"column":0},"generated":{"line":36,"column":0}},{"source":"../node_modules/fbjs/lib/camelizeStyleName.js","original":{"line":37,"column":0},"generated":{"line":37,"column":0}}],"sources":{"../node_modules/fbjs/lib/camelizeStyleName.js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\n'use strict';\n\nvar camelize = require('./camelize');\n\nvar msPattern = /^-ms-/;\n\n/**\n * Camelcases a hyphenated CSS property name, for example:\n *\n *   > camelizeStyleName('background-color')\n *   < \"backgroundColor\"\n *   > camelizeStyleName('-moz-transition')\n *   < \"MozTransition\"\n *   > camelizeStyleName('-ms-transition')\n *   < \"msTransition\"\n *\n * As Andi Smith suggests\n * (http://www.andismith.com/blog/2012/02/modernizr-prefixed/), an `-ms` prefix\n * is converted to lowercase `ms`.\n *\n * @param {string} string\n * @return {string}\n */\nfunction camelizeStyleName(string) {\n  return camelize(string.replace(msPattern, 'ms-'));\n}\n\nmodule.exports = camelizeStyleName;"},"lineCount":37}},"hash":"9b71ea86751e24dfda41d921ff3d7865","cacheData":{"env":{}}}
