+++ website/.cache/3f48b5f19d59f6293a33d96e52a34a49.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/fbjs/lib/camelize.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/package.json","includedInParent":true,"mtime":1535426282286},{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119}],"generated":{"js":"\"use strict\";\n\n/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\nvar _hyphenPattern = /-(.)/g;\n\n/**\n * Camelcases a hyphenated string, for example:\n *\n *   > camelize('background-color')\n *   < \"backgroundColor\"\n *\n * @param {string} string\n * @return {string}\n */\nfunction camelize(string) {\n  return string.replace(_hyphenPattern, function (_, character) {\n    return character.toUpperCase();\n  });\n}\n\nmodule.exports = camelize;","map":{"mappings":[{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":22,"column":0},"generated":{"line":22,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":23,"column":0},"generated":{"line":23,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":24,"column":0},"generated":{"line":24,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":25,"column":0},"generated":{"line":25,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":26,"column":0},"generated":{"line":26,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":27,"column":0},"generated":{"line":27,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":28,"column":0},"generated":{"line":28,"column":0}},{"source":"../node_modules/fbjs/lib/camelize.js","original":{"line":29,"column":0},"generated":{"line":29,"column":0}}],"sources":{"../node_modules/fbjs/lib/camelize.js":"\"use strict\";\n\n/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\nvar _hyphenPattern = /-(.)/g;\n\n/**\n * Camelcases a hyphenated string, for example:\n *\n *   > camelize('background-color')\n *   < \"backgroundColor\"\n *\n * @param {string} string\n * @return {string}\n */\nfunction camelize(string) {\n  return string.replace(_hyphenPattern, function (_, character) {\n    return character.toUpperCase();\n  });\n}\n\nmodule.exports = camelize;"},"lineCount":29}},"hash":"f892443dde660303d0d0150d0ce1a530","cacheData":{"env":{}}}
