+++ website/.cache/5a299fa5b02fde0be5cca6c8d9e3d073.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/fbjs/lib/containsNode.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/package.json","includedInParent":true,"mtime":1535426282286},{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"./isTextNode","loc":{"line":12,"column":25},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/containsNode.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/isTextNode.js"}],"generated":{"js":"'use strict';\n\n/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * \n */\n\nvar isTextNode = require('./isTextNode');\n\n/*eslint-disable no-bitwise */\n\n/**\n * Checks if a given DOM node contains or is another DOM node.\n */\nfunction containsNode(outerNode, innerNode) {\n  if (!outerNode || !innerNode) {\n    return false;\n  } else if (outerNode === innerNode) {\n    return true;\n  } else if (isTextNode(outerNode)) {\n    return false;\n  } else if (isTextNode(innerNode)) {\n    return containsNode(outerNode, innerNode.parentNode);\n  } else if ('contains' in outerNode) {\n    return outerNode.contains(innerNode);\n  } else if (outerNode.compareDocumentPosition) {\n    return !!(outerNode.compareDocumentPosition(innerNode) & 16);\n  } else {\n    return false;\n  }\n}\n\nmodule.exports = containsNode;","map":{"mappings":[{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":22,"column":0},"generated":{"line":22,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":23,"column":0},"generated":{"line":23,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":24,"column":0},"generated":{"line":24,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":25,"column":0},"generated":{"line":25,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":26,"column":0},"generated":{"line":26,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":27,"column":0},"generated":{"line":27,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":28,"column":0},"generated":{"line":28,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":29,"column":0},"generated":{"line":29,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":30,"column":0},"generated":{"line":30,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":31,"column":0},"generated":{"line":31,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":32,"column":0},"generated":{"line":32,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":33,"column":0},"generated":{"line":33,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":34,"column":0},"generated":{"line":34,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":35,"column":0},"generated":{"line":35,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":36,"column":0},"generated":{"line":36,"column":0}},{"source":"../node_modules/fbjs/lib/containsNode.js","original":{"line":37,"column":0},"generated":{"line":37,"column":0}}],"sources":{"../node_modules/fbjs/lib/containsNode.js":"'use strict';\n\n/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * \n */\n\nvar isTextNode = require('./isTextNode');\n\n/*eslint-disable no-bitwise */\n\n/**\n * Checks if a given DOM node contains or is another DOM node.\n */\nfunction containsNode(outerNode, innerNode) {\n  if (!outerNode || !innerNode) {\n    return false;\n  } else if (outerNode === innerNode) {\n    return true;\n  } else if (isTextNode(outerNode)) {\n    return false;\n  } else if (isTextNode(innerNode)) {\n    return containsNode(outerNode, innerNode.parentNode);\n  } else if ('contains' in outerNode) {\n    return outerNode.contains(innerNode);\n  } else if (outerNode.compareDocumentPosition) {\n    return !!(outerNode.compareDocumentPosition(innerNode) & 16);\n  } else {\n    return false;\n  }\n}\n\nmodule.exports = containsNode;"},"lineCount":37}},"hash":"13909f57c6dbdff46e710a19670b5bcb","cacheData":{"env":{}}}
