+++ website/.cache/78b32a1bff071f71c97a5ff479be81da.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"./styles.css","dynamic":true,"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.html","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/styles.css"},{"name":"./index.js","dynamic":true,"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.html","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.js"}],"generated":{"html":"<!DOCTYPE html>\n<html>\n  <head>\n    <meta charset=\"utf-8\">\n    <title>Flictindstergram</title>\n    <link rel=\"stylesheet\" href=\"/19958e0b7edea28fee4313c89dabaf7c.css\">\n  </head>\n  <body>\n    <h1>Hello world</h1>\n    <div id=\"index\"></div>\n    <script src=\"/4c3f856e336061f7fbdd6bce56b32983.js\" type=\"text/javascript\"></script>\n  </body>\n</html>\n"},"hash":"cdcb07e7f18537cb6455812565fbd150","cacheData":{}}
