+++ website/.cache/7a2682c7201e053e864b8971e53a3bab.json
@@ -1,0 +1,1 @@
+{"id":"index.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"react","loc":{"line":1,"column":18},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/react/index.js"},{"name":"react-dom","loc":{"line":2,"column":31},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/react-dom/index.js"},{"name":"./App","loc":{"line":3,"column":16},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/index.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/App.jsx"}],"generated":{"js":"'use strict';\n\nvar _react = require('react');\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactDom = require('react-dom');\n\nvar _App = require('./App');\n\nvar _App2 = _interopRequireDefault(_App);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\n(0, _reactDom.render)(_react2.default.createElement(_App2.default, null), document.getElementById('index'));","map":{"mappings":[{"generated":{"line":3,"column":0},"source":"index.js","original":{"line":1,"column":0}},{"generated":{"line":7,"column":0},"source":"index.js","original":{"line":2,"column":0}},{"generated":{"line":9,"column":0},"source":"index.js","original":{"line":3,"column":0}},{"generated":{"line":15,"column":0},"source":"index.js","original":{"line":4,"column":0}},{"generated":{"line":15,"column":22},"source":"index.js","original":{"line":4,"column":7}},{"generated":{"line":15,"column":52},"source":"index.js","original":{"line":4,"column":8}},{"generated":{"line":15,"column":65},"source":"index.js","original":{"line":4,"column":7}},{"generated":{"line":15,"column":72},"source":"index.js","original":{"line":4,"column":0}},{"name":"document","generated":{"line":15,"column":74},"source":"index.js","original":{"line":4,"column":16}},{"name":"getElementById","generated":{"line":15,"column":83},"source":"index.js","original":{"line":4,"column":25}},{"generated":{"line":15,"column":97},"source":"index.js","original":{"line":4,"column":16}},{"generated":{"line":15,"column":98},"source":"index.js","original":{"line":4,"column":40}},{"generated":{"line":15,"column":105},"source":"index.js","original":{"line":4,"column":16}},{"generated":{"line":15,"column":106},"source":"index.js","original":{"line":4,"column":0}}],"sources":{"index.js":"import React from 'react';\nimport {ReactDOM, render} from 'react-dom';\nimport App from './App';\nrender(<App />, document.getElementById('index'));\n"},"lineCount":null}},"hash":"db7cc6160860cf71201fe9774ecd78a2","cacheData":{"env":{}}}
