+++ website/.cache/7b784b90947238a0a5e2a9205e8a0b14.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/fbjs/lib/emptyObject.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/package.json","includedInParent":true,"mtime":1535426282286},{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119}],"generated":{"js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n */\n\n'use strict';\n\nvar emptyObject = {};\n\nif ('development' !== 'production') {\n  Object.freeze(emptyObject);\n}\n\nmodule.exports = emptyObject;","map":{"mappings":[{"generated":{"line":1,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":1,"column":0}},{"generated":{"line":9,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":9,"column":0}},{"generated":{"line":11,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":11,"column":0}},{"name":"emptyObject","generated":{"line":11,"column":4},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":11,"column":4}},{"generated":{"line":11,"column":18},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":11,"column":18}},{"generated":{"line":11,"column":20},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":11,"column":0}},{"generated":{"line":13,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":13,"column":0}},{"generated":{"line":13,"column":4},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":13,"column":4}},{"generated":{"line":13,"column":22},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":13,"column":29}},{"generated":{"line":13,"column":34},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":13,"column":0}},{"generated":{"line":13,"column":36},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":13,"column":43}},{"name":"Object","generated":{"line":14,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":14,"column":2}},{"name":"freeze","generated":{"line":14,"column":9},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":14,"column":9}},{"generated":{"line":14,"column":15},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":14,"column":2}},{"name":"emptyObject","generated":{"line":14,"column":16},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":14,"column":16}},{"generated":{"line":14,"column":27},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":14,"column":2}},{"generated":{"line":15,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":15,"column":1}},{"name":"module","generated":{"line":17,"column":0},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":17,"column":0}},{"name":"exports","generated":{"line":17,"column":7},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":17,"column":7}},{"generated":{"line":17,"column":14},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":17,"column":0}},{"name":"emptyObject","generated":{"line":17,"column":17},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":17,"column":17}},{"generated":{"line":17,"column":28},"source":"../node_modules/fbjs/lib/emptyObject.js","original":{"line":17,"column":0}}],"sources":{"../node_modules/fbjs/lib/emptyObject.js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n */\n\n'use strict';\n\nvar emptyObject = {};\n\nif (process.env.NODE_ENV !== 'production') {\n  Object.freeze(emptyObject);\n}\n\nmodule.exports = emptyObject;"},"lineCount":null}},"hash":"d49b9271fb9e348b516b9d8fdf056d8c","cacheData":{"env":{"NODE_ENV":"development"}}}
