+++ website/.cache/c892111995aaa753225b417a52f5a964.json
@@ -1,0 +1,1 @@
+{"id":"App.jsx","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"react","loc":{"line":1,"column":18},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/App.jsx","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/react/index.js"}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = require(\"react\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar App = function (_React$Component) {\n    _inherits(App, _React$Component);\n\n    function App(props) {\n        _classCallCheck(this, App);\n\n        var _this = _possibleConstructorReturn(this, (App.__proto__ || Object.getPrototypeOf(App)).call(this, props));\n\n        _this.state = {\n            hello: \"hi\"\n        };\n        return _this;\n    }\n\n    _createClass(App, [{\n        key: \"render\",\n        value: function render() {\n            return _react2.default.createElement(\n                \"div\",\n                null,\n                \"Hello from me\"\n            );\n        }\n    }]);\n\n    return App;\n}(_react2.default.Component);\n\nexports.default = App;","map":{"mappings":[{"generated":{"line":9,"column":0},"source":"App.jsx","original":{"line":1,"column":0}},{"name":"App","generated":{"line":21,"column":4},"source":"App.jsx","original":{"line":3,"column":6}},{"generated":{"line":24,"column":0},"source":"App.jsx","original":{"line":4,"column":4}},{"name":"props","generated":{"line":24,"column":17},"source":"App.jsx","original":{"line":4,"column":16}},{"generated":{"line":24,"column":22},"source":"App.jsx","original":{"line":4,"column":4}},{"generated":{"line":24,"column":24},"source":"App.jsx","original":{"line":4,"column":23}},{"generated":{"line":25,"column":0},"source":"App.jsx","original":{"line":4,"column":23}},{"generated":{"line":27,"column":0},"source":"App.jsx","original":{"line":4,"column":23}},{"name":"props","generated":{"line":27,"column":110},"source":"App.jsx","original":{"line":5,"column":14}},{"generated":{"line":27,"column":115},"source":"App.jsx","original":{"line":4,"column":23}},{"generated":{"line":29,"column":0},"source":"App.jsx","original":{"line":6,"column":8}},{"name":"state","generated":{"line":29,"column":14},"source":"App.jsx","original":{"line":6,"column":13}},{"generated":{"line":29,"column":19},"source":"App.jsx","original":{"line":6,"column":8}},{"generated":{"line":29,"column":22},"source":"App.jsx","original":{"line":6,"column":21}},{"name":"hello","generated":{"line":30,"column":0},"source":"App.jsx","original":{"line":7,"column":12}},{"generated":{"line":30,"column":19},"source":"App.jsx","original":{"line":7,"column":19}},{"generated":{"line":31,"column":0},"source":"App.jsx","original":{"line":6,"column":21}},{"generated":{"line":31,"column":9},"source":"App.jsx","original":{"line":6,"column":8}},{"generated":{"line":32,"column":0},"source":"App.jsx","original":{"line":4,"column":23}},{"generated":{"line":33,"column":0},"source":"App.jsx","original":{"line":9,"column":5}},{"generated":{"line":37,"column":33},"source":"App.jsx","original":{"line":10,"column":14}},{"generated":{"line":38,"column":0},"source":"App.jsx","original":{"line":11,"column":8}},{"generated":{"line":38,"column":19},"source":"App.jsx","original":{"line":12,"column":12}},{"generated":{"line":39,"column":0},"source":"App.jsx","original":{"line":12,"column":12}},{"generated":{"line":40,"column":0},"source":"App.jsx","original":{"line":12,"column":12}},{"generated":{"line":41,"column":0},"source":"App.jsx","original":{"line":12,"column":12}},{"generated":{"line":42,"column":0},"source":"App.jsx","original":{"line":12,"column":12}},{"generated":{"line":42,"column":13},"source":"App.jsx","original":{"line":11,"column":8}},{"generated":{"line":43,"column":0},"source":"App.jsx","original":{"line":14,"column":5}},{"name":"React","generated":{"line":47,"column":2},"source":"App.jsx","original":{"line":3,"column":18}},{"name":"Component","generated":{"line":47,"column":18},"source":"App.jsx","original":{"line":3,"column":24}},{"name":"App","generated":{"line":49,"column":18},"source":"App.jsx","original":{"line":17,"column":15}}],"sources":{"App.jsx":"import React from 'react';\n\nclass App extends React.Component {\n    constructor(props) {\n        super(props)\n        this.state = {\n            hello: \"hi\"\n        }\n    }\n    render () {\n        return (\n            <div>Hello from me</div>\n        )\n    }\n\n} \nexport default App;"},"lineCount":null}},"hash":"5fdc36c4ae89b17185c4216d1964d3e3","cacheData":{"env":{}}}
