+++ website/.cache/f6afe383712fe2185a2c529d6e6fdeb8.json
@@ -1,0 +1,1 @@
+{"id":"styles.css","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"_css_loader","parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/src/styles.css","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/parcel-bundler/src/builtins/css-loader.js"}],"generated":{"css":"body {\n  background-color: blue;\n}","js":"\nvar reloadCSS = require('_css_loader');\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"e4620267164743d91f1e38479e31f2af","cacheData":{"env":{}}}
