+++ website/.cache/f98f61826a86106ae98da7a741f0e17f.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/fbjs/lib/hyphenateStyleName.js","dependencies":[{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/package.json","includedInParent":true,"mtime":1535426282286},{"name":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/package.json","includedInParent":true,"mtime":1535430059119},{"name":"./hyphenate","loc":{"line":12,"column":24},"parent":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/hyphenateStyleName.js","resolved":"/home/andwong91/Codesmith/Projects/Iteration-Project/tinge/website/node_modules/fbjs/lib/hyphenate.js"}],"generated":{"js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\n'use strict';\n\nvar hyphenate = require('./hyphenate');\n\nvar msPattern = /^ms-/;\n\n/**\n * Hyphenates a camelcased CSS property name, for example:\n *\n *   > hyphenateStyleName('backgroundColor')\n *   < \"background-color\"\n *   > hyphenateStyleName('MozTransition')\n *   < \"-moz-transition\"\n *   > hyphenateStyleName('msTransition')\n *   < \"-ms-transition\"\n *\n * As Modernizr suggests (http://modernizr.com/docs/#prefixed), an `ms` prefix\n * is converted to `-ms-`.\n *\n * @param {string} string\n * @return {string}\n */\nfunction hyphenateStyleName(string) {\n  return hyphenate(string).replace(msPattern, '-ms-');\n}\n\nmodule.exports = hyphenateStyleName;","map":{"mappings":[{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":22,"column":0},"generated":{"line":22,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":23,"column":0},"generated":{"line":23,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":24,"column":0},"generated":{"line":24,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":25,"column":0},"generated":{"line":25,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":26,"column":0},"generated":{"line":26,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":27,"column":0},"generated":{"line":27,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":28,"column":0},"generated":{"line":28,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":29,"column":0},"generated":{"line":29,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":30,"column":0},"generated":{"line":30,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":31,"column":0},"generated":{"line":31,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":32,"column":0},"generated":{"line":32,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":33,"column":0},"generated":{"line":33,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":34,"column":0},"generated":{"line":34,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":35,"column":0},"generated":{"line":35,"column":0}},{"source":"../node_modules/fbjs/lib/hyphenateStyleName.js","original":{"line":36,"column":0},"generated":{"line":36,"column":0}}],"sources":{"../node_modules/fbjs/lib/hyphenateStyleName.js":"/**\n * Copyright (c) 2013-present, Facebook, Inc.\n *\n * This source code is licensed under the MIT license found in the\n * LICENSE file in the root directory of this source tree.\n *\n * @typechecks\n */\n\n'use strict';\n\nvar hyphenate = require('./hyphenate');\n\nvar msPattern = /^ms-/;\n\n/**\n * Hyphenates a camelcased CSS property name, for example:\n *\n *   > hyphenateStyleName('backgroundColor')\n *   < \"background-color\"\n *   > hyphenateStyleName('MozTransition')\n *   < \"-moz-transition\"\n *   > hyphenateStyleName('msTransition')\n *   < \"-ms-transition\"\n *\n * As Modernizr suggests (http://modernizr.com/docs/#prefixed), an `ms` prefix\n * is converted to `-ms-`.\n *\n * @param {string} string\n * @return {string}\n */\nfunction hyphenateStyleName(string) {\n  return hyphenate(string).replace(msPattern, '-ms-');\n}\n\nmodule.exports = hyphenateStyleName;"},"lineCount":36}},"hash":"dd109bed9ee9053281053505d9830c78","cacheData":{"env":{}}}
