+++ website/package.json
@@ -4,14 +4,20 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start":"parcel ./src/index.html",
+    "start": "parcel ./src/index.html",
+    "server": "nodemon ./src/server/server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.18.3",
+    "cors": "^2.8.4",
+    "express": "^4.16.3",
     "parcel-bundler": "^1.9.7",
+    "path": "^0.12.7",
     "react": "^16.4.2",
     "react-dom": "^16.4.2"
-  }
+  },
+  "devDependencies": {}
 }
