+++ website/src/index.html
@@ -1,11 +1,13 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
   <head>
     <meta charset="utf-8">
     <title>Flictindstergram</title>
+    <link rel="stylesheet" href="./styles.css">
   </head>
   <body>
+    <h1>Hello world</h1>
     <div id="index"></div>
-    <script src="index.js" type="text/javascript"></script>
+    <script src="./index.js" type="text/javascript"></script>
   </body>
 </html>
