--- website/src/server/app.js
@@ -1,35 +1,0 @@
-const path = require('path');
-const bodyParser = require('body-parser');
-const cors = require('cors');
-const express = require('express');
-const app = express();
-
-const PORT = 3000;
-
-import router from ('./routes.js');
-
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cors());
-
-app.use('/', router);
-
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  var err = new Error('Not Found');
-  err.status = 404;
-  next(err);
-});
-
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
-
-app.listen(PORT);
-// module.exports = app;
