+++ website/src/server/router.js
@@ -1,0 +1,7 @@
+const path = require('path');
+const express = require('express');
+
+// define database calls here
+module.exports = (app) => {
+  
+}
