+++ website/src/server/server.js
@@ -1,0 +1,39 @@
+const path = require('path');
+const bodyParser = require('body-parser');
+const cors = require('cors');
+const express = require('express');
+const app = express();
+const router = require('./router.js');
+
+// localhost port number
+const PORT = 3000;
+
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+
+// allow access from other servers
+app.use(cors());
+
+// define routes before error handlers
+app.use(express.static(path.join(__dirname, './../../dist')));
+router(app);
+
+// catch 404 and forward to error handler
+app.use(function(req, res, next) {
+  var err = new Error('Not Found');
+  err.status = 404;
+  next(err);
+});
+
+// error handler
+app.use(function(err, req, res, next) {
+  // set locals, only providing error in development
+  res.locals.message = err.message;
+
+  // render the error page
+  res.status(err.status || 500);
+  res.send(console.log(err));
+});
+
+app.listen(PORT, () => {console.log(`Listening on port ${PORT}...`)});
+// module.exports = app;
